Corregir manejo de entrada de usuario para evitar NaN en juego de adivinanza

+++ app.js
@@ -2,7 +2,7 @@ const readlineSync = require(`readline-sync`);
 const { generarNumeroAleatorio, verificarAdivinanza } = require("./adivinanza");
 
 const obtenerNumeroUsuario = () => {
-  return parseFloat(readlineSync.question(`Ingrese un numero: `));
+  return parseInt(readlineSync.question(`Ingrese un numero: `));
   
 };
